+++ app/preload/select.js
@@ -35,42 +35,6 @@ ipcRenderer.on("selection-paste", (_, x, y) => documentAtPos(x, y)
     .execCommand("paste"))
 ipcRenderer.on("selection-remove", (_, x, y) => documentAtPos(x, y)
     .getSelection().removeAllRanges())
-
-ipcRenderer.on("download-image-request", (_, x, y) => {
-    const elements = [findElementAtPosition(x, y)]
-    while (elements[0]?.parentNode) {
-        elements.unshift(elements[0].parentNode)
-    }
-    for (const el of elements) {
-        if (el?.tagName?.toLowerCase() === "img" && el.src) {
-            ipcRenderer.sendToHost("download-image", el.src.split("?")[0])
-            break
-        }
-        if (el?.tagName?.toLowerCase() === "svg") {
-            ipcRenderer.sendToHost("download-image", window.URL.createObjectURL(
-                new Blob(el.outerHTML.split(), {"type": "img/svg"})))
-            break
-        }
-        let withUrl = false
-        try {
-            withUrl = getComputedStyle(el).backgroundImage?.startsWith("url")
-        } catch (__) {
-            // Window and top-level nodes don't support getComputedStyle
-        }
-        if (withUrl) {
-            let url = getComputedStyle(el).backgroundImage.slice(4, -1)
-            if (url.startsWith("\"") || url.startsWith("'")) {
-                url = url.slice(1)
-            }
-            if (url.endsWith("\"") || url.endsWith("'")) {
-                url = url.slice(0, -1)
-            }
-            ipcRenderer.sendToHost("download-image", url, true)
-            break
-        }
-    }
-})
-
 ipcRenderer.on("selection-request", (_, startX, startY, endX, endY) => {
     querySelectorAll("*")
     let startNode = findElementAtPosition(startX, startY)